+++ packages/terra-application-layout/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+5.8.0 - (July 23, 2019)
+------------------
 ### Changed
 * Normalized terra-breakpoints dependency version to match other terra-framework packages